chore: Moves CheckboxIcons to Checkbox folder (#13459)
