+++ superset-frontend/src/components/Checkbox/Checkbox.tsx
@@ -16,5 +16,36 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-export { default } from 'src/components/Checkbox/Checkbox';
-export * from 'src/components/Checkbox/CheckboxIcons';
+import React from 'react';
+import { styled } from '@superset-ui/core';
+import { CheckboxChecked, CheckboxUnchecked } from 'src/components/Checkbox';
+
+interface CheckboxProps {
+  checked: boolean;
+  onChange: (val?: boolean) => void;
+  style?: React.CSSProperties;
+}
+
+const Styles = styled.span`
+  &,
+  & svg {
+    vertical-align: top;
+  }
+`;
+
+export default function Checkbox({ checked, onChange, style }: CheckboxProps) {
+  return (
+    <Styles
+      style={style}
+      onClick={() => {
+        onChange(!checked);
+      }}
+      role="checkbox"
+      tabIndex={0}
+      aria-checked={checked}
+      aria-label="Checkbox"
+    >
+      {checked ? <CheckboxChecked /> : <CheckboxUnchecked />}
+    </Styles>
+  );
+}
